refactor: 도메인 별 service 및 repository 위치 이동
