+++ backend/fcm-server/src/main/java/news/bombom/article/service/ArticleArrivalMessageBuilder.java
@@ -1,36 +1,35 @@
-package news.bombom.notification.service.message;
+package news.bombom.article.service;
 
 import java.util.Map;
-import news.bombom.challenge.domain.ChallengeTodoReminderNotification;
+import news.bombom.article.domain.ArticleArrivalNotification;
 import news.bombom.notification.domain.MemberFcmToken;
 import news.bombom.notification.domain.Notification;
 import news.bombom.notification.domain.NotificationPayloadType;
 import news.bombom.notification.domain.NotificationType;
 import news.bombom.notification.dto.NotificationMessage;
+import news.bombom.notification.service.NotificationMessageBuilder;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ChallengeTodoReminderMessageBuilder implements NotificationMessageBuilder {
+public class ArticleArrivalMessageBuilder implements NotificationMessageBuilder {
 
     @Override
     public boolean supports(Notification notification) {
-        return notification instanceof ChallengeTodoReminderNotification;
+        return notification instanceof ArticleArrivalNotification;
     }
 
     @Override
     public NotificationMessage build(Notification notification, MemberFcmToken token) {
-        ChallengeTodoReminderNotification remind = (ChallengeTodoReminderNotification) notification;
-
-        String title = remind.getChallengeName() + " 아직 할 일이 남았어요!";
-        String content = "오늘도 기록을 채워볼까요? 지금 참여해서 꾸준하게 이어가요!";
+        ArticleArrivalNotification article = (ArticleArrivalNotification) notification;
 
         return NotificationMessage.builder()
                 .recipient(token.getFcmToken())
-                .title(title)
-                .content(content)
+                .title(article.getNewsletterName())
+                .content(article.getArticleTitle())
                 .type(NotificationType.FCM)
                 .data(Map.of(
-                        "notificationType", NotificationPayloadType.CHALLENGE_TODO_REMINDER
+                        "articleId", String.valueOf(article.getArticleId()),
+                        "notificationType", NotificationPayloadType.ARTICLE
                 ))
                 .build();
     }
